bookspoiler/requirements.txt
alembic==1.13.1
api==0.0.7
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.7.0
boto3==1.34.88
botocore==1.34.88
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.3
django-allauth==0.61.1
django-restframework==0.0.1
django-storages==1.14.2
djangorestframework==3.15.1
drf-spectacular==0.15.1
environs==11.0.0
filelock==3.13.4
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.1
mysql==0.0.3
mysqlclient==2.2.4
nose==1.3.7
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
platformdirs==4.2.0
plotly==5.20.0
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
s3transfer==0.10.1
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
tenacity==8.2.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.2
waitress==3.0.0
Werkzeug==3.0.2
zope.interface==6.2
# rest_framework_simplejwt 버전...
